Refresher with client reconnect.
The new class MucClient::Refresher can be used to run an independent thread
that periodically checks if a MucClient is still connected, and reconnects
it if not.  In the future, this can also do more things (e.g. clean up
channels that have not been used in a long time).