RPC broadcast server for XMPP broadcast.
This implements a JSON-RPC "broadcast" server that uses the XMPP
broadcast under the hood.  The main logic is encapsulated into a
library class, which allows to run such a server easily.

We will later also add this into a convenience binary, which can be
used locally together with any game-channel daemons to make them XMPP
enabled.