Explicit base64 conversion for RpcServer.
The game-channels RpcBroadcast encodes strings as base64 while
going through the RPC interface.  This change makes the RpcServer
explicitly convert them back to raw before sending via XMPP,
and encode the received messages again before sending them to an
RPC client.

With this, the raw strings (rather than base64-encoded) will be
placed in the XML payload of XMPP stanzas, and base64 is only used
locally when talking to the RpcBroadcast client.

In particular, this ensures that the XMPP protocol is the same
between RpcBroadcast+RpcServer and XmppBroadcast directly, which is
also enforced by a new unit test.